% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clean.R
\name{data_cleaning}
\alias{data_cleaning}
\title{This function selects for certain columns in the AGP data set}
\usage{
data_cleaning(filepath)
}
\arguments{
\item{filepath}{}
}
\value{
data frame with Critter, Mass, SVL, and Stage columns
}
\description{
This function selects for certain columns in the AGP data set
}
\examples{
data_cleaning("/cloud/project/Data/AGP+field+data+copy.csv")
}
