% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hist.R
\name{histogram}
\alias{histogram}
\title{This function creates a histogram of size classes for Amphibians using snout-vent length data}
\usage{
histogram(AGP)
}
\arguments{
\item{AGP}{}
}
\value{
Histogram of data
}
\description{
This function creates a histogram of size classes for Amphibians using snout-vent length data
}
\examples{
histogram(AGP)
}
