% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinMod.R
\name{lin_model}
\alias{lin_model}
\title{This function runs a linear model fit on the provided data}
\usage{
lin_model(filepath)
}
\arguments{
\item{AGP}{}
}
\value{
linear model summary of mass and snout-vent length
}
\description{
This function runs a linear model fit on the provided data
}
\examples{
lin_model(AGP)
}
