% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NoNA.R
\name{nona}
\alias{nona}
\title{This function removes NA values from a given csv file through filepath}
\usage{
nona(filepath)
}
\arguments{
\item{filepath}{}
}
\value{
clean data set
}
\description{
This function removes NA values from a given csv file through filepath
}
\examples{
nona("/cloud/project/Data/AGP+field+data+copy.csv")
}
