% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinModPlot.R
\name{point_plot}
\alias{point_plot}
\title{This function shows the relationship between mass and snout-vent length of amphibians}
\usage{
point_plot(critter)
}
\arguments{
\item{critter}{}
}
\value{
point plot displaying relationship between two variables
}
\description{
This function shows the relationship between mass and snout-vent length of amphibians
}
\examples{
point_plot(AGP)
}
